# docs/api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediKiosk Clinical Intake API",
    "description": "Persistent, clinician-supervised pre-consultation history taking. Not a diagnostic service.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/records/{encounter_id}": {
      "get": {
        "tags": [
          "Clinical records"
        ],
        "summary": "Read Record",
        "operationId": "read_record_records__encounter_id__get",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/records/{encounter_id}/observations": {
      "post": {
        "tags": [
          "Clinical records"
        ],
        "summary": "Add Observation",
        "operationId": "add_observation_records__encounter_id__observations_post",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/records/{encounter_id}/observations/{observation_id}/void": {
      "post": {
        "tags": [
          "Clinical records"
        ],
        "summary": "Void Observation",
        "operationId": "void_observation_records__encounter_id__observations__observation_id__void_post",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          },
          {
            "name": "observation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Observation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/records/{encounter_id}/prescription": {
      "put": {
        "tags": [
          "Clinical records"
        ],
        "summary": "Save Prescription",
        "operationId": "save_prescription_records__encounter_id__prescription_put",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/records/{encounter_id}/pdf": {
      "get": {
        "tags": [
          "Clinical records"
        ],
        "summary": "Get Pdf",
        "operationId": "get_pdf_records__encounter_id__pdf_get",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Medicine": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "strength": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Strength"
          },
          "dose": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Dose"
          },
          "frequency": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Frequency"
          },
          "duration": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Duration"
          },
          "route": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Route"
          },
          "instructions": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Instructions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "strength",
          "dose",
          "frequency",
          "duration",
          "route",
          "instructions"
        ],
        "title": "Medicine"
      },
...
      "ObservationInput": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "VITAL",
              "TEST"
            ],
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "value": {
            "type": "number",
            "maximum": 1000000000.0,
            "minimum": -1000000000.0,
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Unit"
          },
          "measured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Measured At"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "name",
          "value",
          "unit",
          "measured_at",
          "request_id"
        ],
        "title": "ObservationInput"
      },
...
      "PrescriptionInput": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          },
          "medicines": {
            "items": {
              "$ref": "#/components/schemas/Medicine"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Medicines"
          },
          "advice": {
            "type": "string",
            "maxLength": 6000,
            "minLength": 1,
            "title": "Advice"
          },
          "no_medicines_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "No Medicines Reason"
          },
          "follow_up_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up At"
          },
          "allergies_reviewed": {
            "type": "boolean",
            "title": "Allergies Reviewed",
            "default": false
          },
          "allergy_review": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Allergy Review"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "medicines",
          "advice"
        ],
        "title": "PrescriptionInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoidInput": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "VoidInput"
      }
    }
  }
}